Fix comments from review
