+++ packages/slice-machine/src/components/CodeBlock/index.tsx
@@ -6,7 +6,10 @@ import { Text, IconButton } from "@prismicio/editor-ui";
 import theme from "./theme";
 import { Prism as SyntaxHighlighter } from "react-syntax-highlighter";
 
-import * as FileIcons from "@src/icons/FileIcons";
+import { JavaScript } from "@src/icons/JavaScript";
+import { TypeScript } from "@src/icons/TypeScript";
+import { Vue } from "@src/icons/Vue";
+
 import { CodeIcon } from "@src/icons/CodeIcon";
 
 type CodeBlockProps = {
@@ -25,12 +28,12 @@ const FileIcon = ({ fileName }: { fileName: string }) => {
     switch (extension) {
       case "ts":
       case "tsx":
-        return <FileIcons.Typescript />;
+        return <TypeScript />;
       case "js":
       case "jsx":
-        return <FileIcons.Javascript />;
+        return <JavaScript />;
       case "vue":
-        return <FileIcons.Vue />;
+        return <Vue />;
       default:
         return <CodeIcon />;
     }
