+++ packages/slice-machine/src/icons/JavaScript.tsx
@@ -1,0 +1,16 @@
+import { FC } from "react";
+
+export const JavaScript: FC = () => (
+  <svg
+    width="12"
+    height="12"
+    viewBox="0 0 12 12"
+    fill="none"
+    xmlns="http://www.w3.org/2000/svg"
+  >
+    <path
+      d="M11.2388 0H0.762277C0.341518 0 0 0.341518 0 0.762277V11.2377C0 11.6585 0.341518 12 0.762277 12H11.2377C11.6585 12 12 11.6585 12 11.2388V0.762277C12 0.341518 11.6585 0 11.2388 0ZM6.57143 9.19531C6.57143 10.3371 5.90179 10.8571 4.92411 10.8571C4.04018 10.8571 3.4029 10.2645 3.14286 9.71428L4.04129 9.17076C4.21429 9.47768 4.47768 9.71428 4.85714 9.71428C5.21987 9.71428 5.42857 9.57254 5.42857 9.0212V5.42857H6.57143V9.19531ZM9.05022 10.8571C8.03795 10.8571 7.46317 10.3471 7.14286 9.71428L8 9.14286C8.23326 9.52455 8.48772 9.88951 9.02567 9.88951C9.47768 9.88951 9.71428 9.66406 9.71428 9.35156C9.71428 8.97879 9.46875 8.84598 8.9721 8.62835L8.69978 8.51116C7.91295 8.17634 7.38951 7.7567 7.38951 6.8683C7.38951 6.05134 8.01339 5.42857 8.98661 5.42857C9.67969 5.42857 10.1775 5.66964 10.5368 6.30134L9.68862 6.8471C9.50112 6.51116 9.30022 6.37946 8.98661 6.37946C8.66741 6.37946 8.4654 6.58259 8.4654 6.8471C8.4654 7.17411 8.66741 7.3058 9.13504 7.50893L9.40737 7.62612C10.3348 8.02232 10.8571 8.42745 10.8571 9.33817C10.8571 10.3192 10.0859 10.8571 9.05022 10.8571Z"
+      fill="#6F6E77"
+    />
+  </svg>
+);
