+++ packages/slice-machine/src/icons/TypeScript.tsx
@@ -1,6 +1,6 @@
 import { FC } from "react";
 
-export const Typescript: FC = () => (
+export const TypeScript: FC = () => (
   <svg
     width="12"
     height="12"
@@ -12,38 +12,6 @@ export const Typescript: FC = () => (
       fillRule="evenodd"
       clipRule="evenodd"
       d="M1.17188 0H10.8281C11.4753 0 12 0.524666 12 1.17188V10.8281C12 11.4753 11.4753 12 10.8281 12H1.17188C0.524666 12 0 11.4753 0 10.8281V1.17188C0 0.524666 0.524666 0 1.17188 0ZM7.42826 10.7223V9.549C7.641 9.72745 7.87209 9.86128 8.12152 9.95048C8.37096 10.0397 8.62282 10.0843 8.87714 10.0843C9.02632 10.0843 9.15652 10.0709 9.2678 10.044C9.37906 10.0171 9.47198 9.97983 9.54656 9.93216C9.62114 9.88448 9.67678 9.82828 9.71346 9.7635C9.75014 9.69872 9.76849 9.62845 9.76849 9.55268C9.76849 9.45 9.73915 9.35834 9.68046 9.27769C9.62175 9.19702 9.54166 9.12246 9.44018 9.05402C9.33872 8.98559 9.21827 8.91959 9.07889 8.85602C8.93951 8.79246 8.78911 8.7277 8.62772 8.6617C8.21688 8.49059 7.9106 8.28159 7.70887 8.03473C7.50712 7.78784 7.40625 7.48962 7.40625 7.14007C7.40625 6.8663 7.46128 6.63103 7.57132 6.43427C7.68136 6.23749 7.83112 6.07556 8.02064 5.94844C8.21018 5.82134 8.42965 5.72784 8.67907 5.66796C8.92849 5.60808 9.19259 5.57812 9.47137 5.57812C9.74524 5.57812 9.98796 5.59462 10.1995 5.62762C10.411 5.66062 10.606 5.71134 10.7845 5.77978V6.87609C10.6965 6.81497 10.6005 6.7612 10.4966 6.71475C10.3927 6.66832 10.2857 6.62981 10.1756 6.59925C10.0656 6.56871 9.95616 6.54609 9.84734 6.53142C9.73852 6.51677 9.6352 6.50944 9.5374 6.50944C9.40289 6.50944 9.28064 6.52226 9.17058 6.54792C9.06054 6.57359 8.96763 6.60966 8.89181 6.65609C8.81602 6.70254 8.75733 6.75813 8.71575 6.82291C8.67417 6.8877 8.65338 6.96042 8.65338 7.04107C8.65338 7.12908 8.67663 7.2079 8.72309 7.27758C8.76954 7.34723 8.83556 7.41323 8.92116 7.47558C9.00675 7.5379 9.11067 7.599 9.23295 7.65891C9.35522 7.71879 9.49338 7.7805 9.64744 7.84406C9.85774 7.93205 10.0466 8.02554 10.2142 8.12454C10.3817 8.22354 10.5253 8.33538 10.6451 8.46005C10.765 8.58471 10.8567 8.72709 10.9203 8.8872C10.9838 9.0473 11.0156 9.23367 11.0156 9.44634C11.0156 9.73966 10.96 9.98592 10.8487 10.1851C10.7375 10.3844 10.5865 10.5457 10.3957 10.6691C10.205 10.7926 9.98306 10.8812 9.72996 10.935C9.47686 10.9887 9.20972 11.0156 8.92849 11.0156C8.63993 11.0156 8.36545 10.9912 8.10502 10.9423C7.84458 10.8934 7.61899 10.8201 7.42826 10.7223ZM6.96094 6.65817H5.45578V10.9453H4.26342V6.65817H2.76562V5.69531H6.96094V6.65817Z"
-      fill="#6F6E77"
-    />
-  </svg>
-);
-
-export const Javascript: FC = () => (
-  <svg
-    width="12"
-    height="12"
-    viewBox="0 0 12 12"
-    fill="none"
-    xmlns="http://www.w3.org/2000/svg"
-  >
-    <path
-      d="M11.2388 0H0.762277C0.341518 0 0 0.341518 0 0.762277V11.2377C0 11.6585 0.341518 12 0.762277 12H11.2377C11.6585 12 12 11.6585 12 11.2388V0.762277C12 0.341518 11.6585 0 11.2388 0ZM6.57143 9.19531C6.57143 10.3371 5.90179 10.8571 4.92411 10.8571C4.04018 10.8571 3.4029 10.2645 3.14286 9.71428L4.04129 9.17076C4.21429 9.47768 4.47768 9.71428 4.85714 9.71428C5.21987 9.71428 5.42857 9.57254 5.42857 9.0212V5.42857H6.57143V9.19531ZM9.05022 10.8571C8.03795 10.8571 7.46317 10.3471 7.14286 9.71428L8 9.14286C8.23326 9.52455 8.48772 9.88951 9.02567 9.88951C9.47768 9.88951 9.71428 9.66406 9.71428 9.35156C9.71428 8.97879 9.46875 8.84598 8.9721 8.62835L8.69978 8.51116C7.91295 8.17634 7.38951 7.7567 7.38951 6.8683C7.38951 6.05134 8.01339 5.42857 8.98661 5.42857C9.67969 5.42857 10.1775 5.66964 10.5368 6.30134L9.68862 6.8471C9.50112 6.51116 9.30022 6.37946 8.98661 6.37946C8.66741 6.37946 8.4654 6.58259 8.4654 6.8471C8.4654 7.17411 8.66741 7.3058 9.13504 7.50893L9.40737 7.62612C10.3348 8.02232 10.8571 8.42745 10.8571 9.33817C10.8571 10.3192 10.0859 10.8571 9.05022 10.8571Z"
-      fill="#6F6E77"
-    />
-  </svg>
-);
-
-export const Vue: FC = () => (
-  <svg
-    width="12"
-    height="11"
-    viewBox="0 0 12 11"
-    fill="none"
-    xmlns="http://www.w3.org/2000/svg"
-  >
-    <path
-      fillRule="evenodd"
-      clipRule="evenodd"
-      d="M2.03097 0L5.75841 6.83364L6.08306 7.42884L6.40772 6.83364L10.1352 0H12L6.08306 10.4347L0 0H2.03097ZM9.29235 0.000579537L6.08306 5.88427L2.87373 0.00049377L4.26119 0.000124105L4.72699 0L5.27775 0.949738L6.13189 2.42265L6.98856 0.951205L7.54201 0.000579537H7.98814H9.29235Z"
       fill="#6F6E77"
     />
   </svg>
