+++ packages/slice-machine/src/icons/Vue.tsx
@@ -1,0 +1,18 @@
+import { FC } from "react";
+
+export const Vue: FC = () => (
+  <svg
+    width="12"
+    height="11"
+    viewBox="0 0 12 11"
+    fill="none"
+    xmlns="http://www.w3.org/2000/svg"
+  >
+    <path
+      fillRule="evenodd"
+      clipRule="evenodd"
+      d="M2.03097 0L5.75841 6.83364L6.08306 7.42884L6.40772 6.83364L10.1352 0H12L6.08306 10.4347L0 0H2.03097ZM9.29235 0.000579537L6.08306 5.88427L2.87373 0.00049377L4.26119 0.000124105L4.72699 0L5.27775 0.949738L6.13189 2.42265L6.98856 0.951205L7.54201 0.000579537H7.98814H9.29235Z"
+      fill="#6F6E77"
+    />
+  </svg>
+);
